Split off Events from core/backbone
